SKFP-143: Add express middleware to resolve sets id in sqon and remove search endpoint
