--- src/endpoints/search.test.ts
@@ -1,157 +1,0 @@
-import { resolveSetsInSqon } from '../sqon/resolveSetInSqon';
-import { ArrangerProject } from '../sqon/searchSqon';
-import { search, SearchVariables } from './search';
-
-jest.mock('../sqon/resolveSetInSqon');
-
-describe('Search feature', () => {
-    const projectId = '2021_05_03_v2';
-    const userId = 'userId';
-    const accessToken = 'access_token';
-    const variables: SearchVariables = {
-        sqon: {
-            op: 'and',
-            content: [
-                {
-                    op: 'in',
-                    content: {
-                        field: 'kf_id',
-                        value: ['set_id:abc'],
-                    },
-                },
-            ],
-        },
-    };
-    const query = 'query($sqon: JSON) {participant {hits(filters: $sqon) {total}}}';
-
-    beforeEach(() => {
-        (resolveSetsInSqon as jest.Mock).mockReset();
-    });
-
-    it('should replace setId by participants list and send query to Arranger', async () => {
-        const expectedSqon = {
-            ...variables.sqon,
-            content: [
-                {
-                    op: 'in',
-                    content: {
-                        field: 'kf_id',
-                        value: ['participant_1', 'participant_2'],
-                    },
-                },
-            ],
-        };
-
-        const expectedSearchResult = {
-            data: {
-                participant: {
-                    hits: {
-                        total: 198,
-                    },
-                },
-            },
-        };
-
-        (resolveSetsInSqon as jest.Mock).mockImplementation(() => expectedSqon);
-
-        const project: ArrangerProject = {
-            runQuery: () => Promise.resolve(expectedSearchResult),
-        };
-
-        const result = await search(userId, accessToken, projectId, query, variables, () => project);
-
-        expect(result).toEqual(expectedSearchResult);
-        expect((resolveSetsInSqon as jest.Mock).mock.calls.length).toEqual(1);
-    });
-
-    it('should send query to Arranger directly if no variables or no sqon in variables', async () => {
-        const expectedSearchResult = {
-            data: {
-                participant: {
-                    hits: {
-                        total: 198,
-                    },
-                },
-            },
-        };
-
-        const project: ArrangerProject = {
-            runQuery: () => Promise.resolve(expectedSearchResult),
-        };
-
-        const resultVariablesNull = await search(userId, accessToken, projectId, query, null, () => project);
-        expect(resultVariablesNull).toEqual(expectedSearchResult);
-        const resultVariablesUndefined = await search(userId, accessToken, projectId, query, undefined, () => project);
-        expect(resultVariablesUndefined).toEqual(expectedSearchResult);
-        const resultSqonNull = await search(userId, accessToken, projectId, query, { sqon: null }, () => project);
-        expect(resultSqonNull).toEqual(expectedSearchResult);
-        const resultSqonUndefined = await search(
-            userId,
-            accessToken,
-            projectId,
-            query,
-            { sqon: undefined },
-            () => project,
-        );
-        expect(resultSqonUndefined).toEqual(expectedSearchResult);
-
-        expect((resolveSetsInSqon as jest.Mock).mock.calls.length).toEqual(0);
-    });
-
-    it('should throw an error if replacing setId failed', async () => {
-        const expectedSearchResult = {
-            data: {
-                participant: {
-                    hits: {
-                        total: 198,
-                    },
-                },
-            },
-        };
-
-        (resolveSetsInSqon as jest.Mock).mockImplementation(() => {
-            throw new Error('OOPS');
-        });
-
-        const project: ArrangerProject = {
-            runQuery: () => Promise.resolve(expectedSearchResult),
-        };
-
-        try {
-            await search(userId, accessToken, projectId, query, variables, () => project);
-        } catch (e) {
-            expect(e.message).toEqual('OOPS');
-        } finally {
-            expect((resolveSetsInSqon as jest.Mock).mock.calls.length).toEqual(1);
-        }
-    });
-
-    it('should throw an error if project does not exist', async () => {
-        const expectedSqon = {
-            ...variables.sqon,
-            content: [
-                {
-                    op: 'in',
-                    content: {
-                        field: 'kf_id',
-                        value: ['participant_1', 'participant_2'],
-                    },
-                },
-            ],
-        };
-
-        (resolveSetsInSqon as jest.Mock).mockImplementation(() => expectedSqon);
-
-        const project: ArrangerProject = {
-            runQuery: () => Promise.resolve(undefined),
-        };
-
-        try {
-            await search(userId, accessToken, projectId, query, variables, () => project);
-        } catch (e) {
-            expect(e.message).toEqual(`ProjectID '${projectId}' cannot be established.`);
-        } finally {
-            expect((resolveSetsInSqon as jest.Mock).mock.calls.length).toEqual(1);
-        }
-    });
-});
